+++ NEWS.md
@@ -6,6 +6,7 @@
 ### Fixed
 - Fixed display problems with Choose R dialog when UI language is French #12717
 -
+- Background script jobs are now run using the global environment. This fixes the behaviour of `source()` in backgrounds jobs. (#11866)
 
 ### Accessibility Improvements
 -